refactor: examples renaming to get rid of bad use of capital letters
