+++ examples/adiabaticparcel/src/CMakeLists.txt
@@ -12,18 +12,18 @@ message("PROJECT_SOURCE_DIR: ${PROJECT_SOURCE_DIR}")
 set(CLEOLIBS gridboxes initialise observers runcleo superdrops zarr)
 
 # create primary executable for CVODE-CLEO coupled model adiabatic parcel setup
-add_executable(adia0D EXCLUDE_FROM_ALL "main_adia0D.cpp")
+add_executable(adia0d EXCLUDE_FROM_ALL "main_adia0d.cpp")
 
 # Add directories and link libraries to target
-target_link_libraries(adia0D PRIVATE coupldyn_cvode cartesiandomain "${CLEOLIBS}")
-target_link_libraries(adia0D PUBLIC Kokkos::kokkos)
-target_include_directories(adia0D PRIVATE "${CLEO_SOURCE_DIR}/libs") # CLEO libs directory
+target_link_libraries(adia0d PRIVATE coupldyn_cvode cartesiandomain "${CLEOLIBS}")
+target_link_libraries(adia0d PUBLIC Kokkos::kokkos)
+target_include_directories(adia0d PRIVATE "${CLEO_SOURCE_DIR}/libs") # CLEO libs directory
 
 # set specific C++ compiler options for target (optional)
-#target_compile_options(adia0D PRIVATE)
+#target_compile_options(adia0d PRIVATE)
 
 # set compiler properties for target(s)
-set_target_properties(adia0D PROPERTIES
+set_target_properties(adia0d PROPERTIES
   CMAKE_CXX_STANDARD_REQUIRED ON
   CMAKE_CXX_EXTENSIONS ON
   CXX_STANDARD 20)
