+++ examples/boxmodelcollisions/breakup_config.yaml
@@ -46,9 +46,9 @@ initsupers:
 
 ### Output Parameters ###
 outputdata:
-  setup_filename: /home/m/m300950/CLEO/build_colls0D/bin/testikstraub_setup.txt  # .txt filename to copy configuration to
-  stats_filename: /home/m/m300950/CLEO/build_colls0D/bin/testikstraub_stats.txt  # .txt file to output runtime statistics to
-  zarrbasedir: /home/m/m300950/CLEO/build_colls0D/bin/testikstraub_sol.zarr  # zarr store base directory
+  setup_filename: /home/m/m300950/CLEO/build_colls0d/bin/testikstraub_setup.txt  # .txt filename to copy configuration to
+  stats_filename: /home/m/m300950/CLEO/build_colls0d/bin/testikstraub_stats.txt  # .txt file to output runtime statistics to
+  zarrbasedir: /home/m/m300950/CLEO/build_colls0d/bin/testikstraub_sol.zarr  # zarr store base directory
   maxchunk: 2500000                                           # maximum no. of elements in chunks of zarr store array
 
 ### Microphysics Parameters ###
