+++ examples/bubble3d/src/CMakeLists.txt
@@ -12,18 +12,18 @@ message("PROJECT_SOURCE_DIR: ${PROJECT_SOURCE_DIR}")
 set(CLEOLIBS gridboxes initialise observers runcleo superdrops zarr)
 
 # create primary executable for CLEO in 2-D coupled to thermodynamics from file setup
-add_executable(bubble3D EXCLUDE_FROM_ALL "main_bubble3d.cpp")
+add_executable(bubble3d EXCLUDE_FROM_ALL "main_bubble3d.cpp")
 
 # Add directories and link libraries to target
-target_link_libraries(bubble3D PRIVATE coupldyn_yac cartesiandomain "${CLEOLIBS}")
-target_link_libraries(bubble3D PUBLIC Kokkos::kokkos)
-target_include_directories(bubble3D PRIVATE "${CLEO_SOURCE_DIR}/libs") # CLEO libs directory
+target_link_libraries(bubble3d PRIVATE coupldyn_yac cartesiandomain "${CLEOLIBS}")
+target_link_libraries(bubble3d PUBLIC Kokkos::kokkos)
+target_include_directories(bubble3d PRIVATE "${CLEO_SOURCE_DIR}/libs") # CLEO libs directory
 
 # set specific C++ compiler options for target (optional)
-#target_compile_options(bubble3D PRIVATE)
+#target_compile_options(bubble3d PRIVATE)
 
 # set compiler properties for target(s)
-set_target_properties(bubble3D PROPERTIES
+set_target_properties(bubble3d PROPERTIES
   CMAKE_CXX_STANDARD_REQUIRED ON
   CMAKE_CXX_EXTENSIONS ON
   CXX_STANDARD 20)
