+++ examples/constthermo2d/constthermo2d.sh
@@ -20,9 +20,9 @@
 ### ---------------------------------------------------- ###
 buildtype="cuda"
 path2CLEO=${HOME}/CLEO/
-path2build=${HOME}/CLEO/build_const2D/
+path2build=${HOME}/CLEO/build_const2d/
 enableyac=false
-executables="const2D"
+executables="const2d"
 
 pythonscript=${path2CLEO}/examples/constthermo2d/constthermo2d.py
 configfile=${path2CLEO}/examples/constthermo2d/src/config/const2d_config.yaml
