+++ examples/divfreemotion/src/main_divfree2d.cpp
@@ -3,26 +3,25 @@
  *
  *
  * ----- CLEO -----
- * File: main_eurec4a1D.cpp
+ * File: main_divfree2d.cpp
  * Project: src
- * Created Date: Tuesday 9th April 2024
+ * Created Date: Monday 29th January 2024
  * Author: Clara Bayley (CB)
  * Additional Contributors:
  * -----
- * Last Modified: Tuesday 18th June 2024
+ * Last Modified: Wednesday 11th September 2024
  * Modified By: CB
  * -----
  * License: BSD 3-Clause "New" or "Revised" License
  * https://opensource.org/licenses/BSD-3-Clause
  * -----
  * File Description:
- * runs the CLEO super-droplet model (SDM) for eurec4a 1-D rainshaft example.
- * After make/compiling, execute for example via:
- * ./src/eurec4a1D ../src/config/config.yaml
+ * runs the CLEO super-droplet model (SDM) for 2-D divergence free motion example
+ * after make/compiling, execute for example via:
+ * ./src/divfree2d ../src/config/config.yaml
  */
 
 #include <Kokkos_Core.hpp>
-#include <array>
 #include <cmath>
 #include <concepts>
 #include <iostream>
@@ -30,7 +29,6 @@
 #include <string_view>
 
 #include "zarr/dataset.hpp"
-#include "cartesiandomain/add_supers_at_domain_top.hpp"
 #include "cartesiandomain/cartesianmaps.hpp"
 #include "cartesiandomain/cartesianmotion.hpp"
 #include "cartesiandomain/createcartesianmaps.hpp"
@@ -40,31 +38,20 @@
 #include "gridboxes/gridboxmaps.hpp"
 #include "initialise/config.hpp"
 #include "initialise/init_all_supers_from_binary.hpp"
-#include "initialise/init_supers_from_binary.hpp"
 #include "initialise/initgbxsnull.hpp"
 #include "initialise/initialconditions.hpp"
 #include "initialise/timesteps.hpp"
 #include "observers/gbxindex_observer.hpp"
-#include "observers/massmoments_observer.hpp"
-#include "observers/nsupers_observer.hpp"
 #include "observers/observers.hpp"
-#include "observers/runstats_observer.hpp"
-#include "observers/sdmmonitor/monitor_condensation_observer.hpp"
-#include "observers/state_observer.hpp"
 #include "observers/streamout_observer.hpp"
 #include "observers/superdrops_observer.hpp"
 #include "observers/time_observer.hpp"
-#include "observers/windvel_observer.hpp"
 #include "runcleo/coupleddynamics.hpp"
 #include "runcleo/couplingcomms.hpp"
 #include "runcleo/runcleo.hpp"
 #include "runcleo/sdmmethods.hpp"
-#include "superdrops/collisions/coalescence.hpp"
-#include "superdrops/collisions/longhydroprob.hpp"
-#include "superdrops/condensation.hpp"
 #include "superdrops/microphysicalprocess.hpp"
 #include "superdrops/motion.hpp"
-#include "superdrops/terminalvelocity.hpp"
 #include "zarr/fsstore.hpp"
 
 inline CoupledDynamics auto create_coupldyn(const Config &config, const CartesianMaps &gbxmaps,
@@ -78,9 +65,9 @@ inline CoupledDynamics auto create_coupldyn(const Config &config, const Cartesia
   return FromFileDynamics(config.get_fromfiledynamics(), couplstep, ndims, nsteps);
 }
 
-inline InitialConditions auto create_initconds(const Config &config, const CartesianMaps &gbxmaps) {
-  const InitGbxsNull initgbxs(gbxmaps.get_total_local_gridboxes());
-  const InitSupersFromBinary initsupers(config.get_initsupersfrombinary(), gbxmaps);
+inline InitialConditions auto create_initconds(const Config &config) {
+  const InitAllSupersFromBinary initsupers(config.get_initsupersfrombinary());
+  const InitGbxsNull initgbxs(config.get_ngbxs());
 
   return InitConds(initsupers, initgbxs);
 }
@@ -91,29 +78,19 @@ inline GridboxMaps auto create_gbxmaps(const Config &config) {
   return gbxmaps;
 }
 
-inline auto create_movement(const Config &config, const Timesteps &tsteps,
-                            const CartesianMaps &gbxmaps) {
-  const auto terminalv = RogersGKTerminalVelocity{};
+inline auto create_movement(const unsigned int motionstep, const CartesianMaps &gbxmaps) {
+  const auto terminalv = NullTerminalVelocity{};
   const Motion<CartesianMaps> auto motion =
-      CartesianMotion(tsteps.get_motionstep(), &step2dimlesstime, terminalv);
+      CartesianMotion(motionstep, &step2dimlesstime, terminalv);
 
-  // const auto boundary_conditions = NullBoundaryConditions{};
-  const auto boundary_conditions = AddSupersAtDomainTop(config.get_addsupersatdomaintop());
+  const auto boundary_conditions = NullBoundaryConditions{};
 
   return MoveSupersInDomain(gbxmaps, motion, boundary_conditions);
 }
 
 inline MicrophysicalProcess auto create_microphysics(const Config &config,
                                                      const Timesteps &tsteps) {
-  const auto c = config.get_condensation();
-  const MicrophysicalProcess auto cond =
-      Condensation(tsteps.get_condstep(), &step2dimlesstime, c.do_alter_thermo, c.maxniters, c.rtol,
-                   c.atol, c.MINSUBTSTEP, &realtime2dimless);
-
-  const PairProbability auto coalprob = LongHydroProb(1.0);
-  const MicrophysicalProcess auto coal = CollCoal(tsteps.get_collstep(), &step2realtime, coalprob);
-
-  return coal >> cond;
+  return NullMicrophysicalProcess{};
 }
 
 template <typename Store>
@@ -121,26 +98,11 @@ inline Observer auto create_superdrops_observer(const unsigned int interval,
                                                 Dataset<Store> &dataset, const int maxchunk) {
   CollectDataForDataset<Store> auto sdid = CollectSdId(dataset, maxchunk);
   CollectDataForDataset<Store> auto sdgbxindex = CollectSdgbxindex(dataset, maxchunk);
-  CollectDataForDataset<Store> auto xi = CollectXi(dataset, maxchunk);
-  CollectDataForDataset<Store> auto radius = CollectRadius(dataset, maxchunk);
-  CollectDataForDataset<Store> auto msol = CollectMsol(dataset, maxchunk);
   CollectDataForDataset<Store> auto coord3 = CollectCoord3(dataset, maxchunk);
+  CollectDataForDataset<Store> auto coord1 = CollectCoord1(dataset, maxchunk);
 
-  const auto collect_sddata = coord3 >> msol >> radius >> xi >> sdgbxindex >> sdid;
+  const auto collect_sddata = coord1 >> coord3 >> sdgbxindex >> sdid;
   return SuperdropsObserver(interval, dataset, maxchunk, collect_sddata);
-}
-
-template <typename Store>
-inline Observer auto create_gridboxes_observer(const unsigned int interval, Dataset<Store> &dataset,
-                                               const int maxchunk, const size_t ngbxs) {
-  const CollectDataForDataset<Store> auto thermo = CollectThermo(dataset, maxchunk, ngbxs);
-  const CollectDataForDataset<Store> auto wvel =
-      CollectWindVariable<Store, WvelFunc>(dataset, WvelFunc{}, "wvel", maxchunk, ngbxs);
-
-  const CollectDataForDataset<Store> auto nsupers = CollectNsupers(dataset, maxchunk, ngbxs);
-
-  const CollectDataForDataset<Store> auto collect_gbxdata = nsupers >> wvel >> thermo;
-  return WriteToDatasetObserver(interval, dataset, collect_gbxdata);
 }
 
 template <typename Store>
@@ -148,27 +110,14 @@ inline Observer auto create_observer(const Config &config, const Timesteps &tste
                                      Dataset<Store> &dataset) {
   const auto obsstep = tsteps.get_obsstep();
   const auto maxchunk = config.get_maxchunk();
-  const auto ngbxs = config.get_ngbxs();
 
-  const Observer auto obs0 = RunStatsObserver(obsstep, config.get_stats_filename());
+  const Observer auto obs0 = StreamOutObserver(obsstep, &step2realtime);
 
-  const Observer auto obs1 = StreamOutObserver(realtime2step(240), &step2realtime);
-
-  const Observer auto obs2 = TimeObserver(obsstep, dataset, maxchunk, &step2dimlesstime);
-
-  const Observer auto obs3 = GbxindexObserver(dataset, maxchunk, ngbxs);
-
-  const Observer auto obs5 = MassMomentsObserver(obsstep, dataset, maxchunk, ngbxs);
-
-  const Observer auto obs6 = MassMomentsRaindropsObserver(obsstep, dataset, maxchunk, ngbxs);
-
-  const Observer auto obsgbx = create_gridboxes_observer(obsstep, dataset, maxchunk, ngbxs);
+  const Observer auto obs1 = TimeObserver(obsstep, dataset, maxchunk, &step2dimlesstime);
 
   const Observer auto obssd = create_superdrops_observer(obsstep, dataset, maxchunk);
 
-  const Observer auto obs_cond = MonitorCondensationObserver(obsstep, dataset, maxchunk, ngbxs);
-
-  return obs_cond >> obssd >> obsgbx >> obs6 >> obs5 >> obs3 >> obs2 >> obs1 >> obs0;
+  return obssd >> obs1 >> obs0;
 }
 
 template <typename Store>
@@ -176,7 +125,7 @@ inline auto create_sdm(const Config &config, const Timesteps &tsteps, Dataset<St
   const auto couplstep = (unsigned int)tsteps.get_couplstep();
   const GridboxMaps auto gbxmaps(create_gbxmaps(config));
   const MicrophysicalProcess auto microphys(create_microphysics(config, tsteps));
-  const MoveSupersInDomain movesupers(create_movement(config, tsteps, gbxmaps));
+  const MoveSupersInDomain movesupers(create_movement(tsteps.get_motionstep(), gbxmaps));
   const Observer auto obs(create_observer(config, tsteps, dataset));
 
   return SDMMethods(couplstep, gbxmaps, microphys, movesupers, obs);
@@ -208,14 +157,14 @@ int main(int argc, char *argv[]) {
   auto store = FSStore(config.get_zarrbasedir());
   auto dataset = Dataset(store);
 
+  /* Initial conditions for CLEO run */
+  const InitialConditions auto initconds = create_initconds(config);
+
   /* Initialise Kokkos parallel environment */
   Kokkos::initialize(argc, argv);
   {
     /* CLEO Super-Droplet Model (excluding coupled dynamics solver) */
     const SDMMethods sdm(create_sdm(config, tsteps, dataset));
-
-    /* Initial conditions for CLEO run */
-    const InitialConditions auto initconds = create_initconds(config, sdm.gbxmaps);
 
     /* Solver of dynamics coupled to CLEO SDM */
     CoupledDynamics auto coupldyn(
