+++ examples/eurec4a1d/eurec4a1d.sh
@@ -21,9 +21,9 @@
 ### ---------------------------------------------------- ###
 buildtype="cuda"
 path2CLEO=${HOME}/CLEO/
-path2build=${HOME}/CLEO/build_eurec4a1D/
+path2build=${HOME}/CLEO/build_eurec4a1d/
 enableyac=false
-executables="eurec4a1D"
+executables="eurec4a1d"
 
 configfile=${path2CLEO}/examples/eurec4a1d/src/config/eurec4a1d_config.yaml
 pythonscript=""
